+++ r-logs.tf
@@ -1,53 +1,10 @@
-resource "azurerm_monitor_diagnostic_setting" "log_settings_storage" {
-  count = var.enable_logs_to_storage ? 1 : 0
+module "diagnostics" {
+  source  = "claranet/diagnostic-settings/azurerm"
+  version = "4.0.1"
 
-  name               = "logs-storage"
-  target_resource_id = azurerm_key_vault.keyvault.id
-
-  storage_account_id = var.logs_storage_account_id
-
-  log {
-    category = "AuditEvent"
-
-    retention_policy {
-      enabled = true
-      days    = var.logs_storage_retention
-    }
-  }
-
-  metric {
-    category = "AllMetrics"
-    enabled  = false
-
-    retention_policy {
-      enabled = false
-    }
-  }
+  resource_id           = azurerm_key_vault.keyvault.id
+  logs_destinations_ids = var.logs_destinations_ids
+  log_categories        = var.logs_categories
+  metric_categories     = var.logs_metrics_categories
+  retention_days        = var.logs_retention_days
 }
-
-resource "azurerm_monitor_diagnostic_setting" "log_settings_log_analytics" {
-  count = var.enable_logs_to_log_analytics ? 1 : 0
-
-  name               = "logs-log-analytics"
-  target_resource_id = azurerm_key_vault.keyvault.id
-
-  log_analytics_workspace_id = var.logs_log_analytics_workspace_id
-
-  log {
-    category = "AuditEvent"
-
-    retention_policy {
-      enabled = false
-    }
-  }
-
-  metric {
-    category = "AllMetrics"
-    enabled  = false
-
-    retention_policy {
-      enabled = false
-    }
-  }
-}
-