+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Breaking
   * AZ-482: `soft_delete_enabled` is now deprecated since provider [v2.42.0](https://github.com/terraform-providers/terraform-provider-azurerm/blob/master/CHANGELOG-v2.md#2420-january-08-2021) and should be removed from this module.
+  * AZ-160: Unify diagnostics settings on all Claranet modules
 
 # v4.2.0 - 2021-03-09